+++ code/app/routes.py
@@ -5,12 +5,6 @@ from wtforms import StringField, PasswordField, BooleanField, SubmitField
 from wtforms.validators import NumberRange
 from flask_login import login_user, logout_user, current_user, login_required
 
-class AddToCartButton(FlaskForm):
-    cartsubmit = SubmitField('Add to cart')
-
-class RateItemButton(FlaskForm):
-    rating = IntegerField('Rating (1-10):', validators=[NumberRange(min=1, max=10)])
-    ratingsubmit = SubmitField('Rate') 
 
 @myapp_obj.route('/')
 @myapp_obj.route('/login')
